# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction",
        "description": "Documentation Auction web ",
        "contact": {
            "email": "rustamovvramziddin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Post new University  data",
                "operationId": "3baa3b145384161275c3581002c9225f",
                "requestBody": {
                    "description": "Pass Auth   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Post new user data",
                "operationId": "9f2c210e9e11a13335b03c7e02a95215",
                "requestBody": {
                    "description": "Pass Auth   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "fullname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Doe"
                                    },
                                    "fullname": {
                                        "type": "text",
                                        "format": "",
                                        "example": "Surname First name"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    },
                                    "confirmation_password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Logout  data",
                "operationId": "0c6d22e05b701dd1f04b4421293a432f",
                "requestBody": {
                    "description": "Logout ",
                    "required": true
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "api": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}